requirements.txt
Django==5.0.4
python-dotenv==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.4
django-celery-beat==2.6.0
requests==2.31.0
coverage==7.5.1
django-cors-headers==4.3.1
flake8==7.0.0

